Pass custom escape function to included templates
At the moment, if you have a template that includes another template,
and a custom escape function, the escape function is not correctly
passed from the outer template to the included template. For an
example of this, see the test fixture added in this commit, which
failed before this fix.

This is due to the property being renamed from `options.escape` to
`options.escapeFunction` in the `Template` constructor. Then, when a
nested `Template` is created, it tries to look for `options.escape`, but
the option has been renamed to `options.escapeFunction`.

This change replaces use of `options.escapeFunction` to `options.escape`
in order to correctly pass the escape function down to included
templates.